(Refactor) refactored API to have a routes directory, removed src directory
