+++ api/app.py
@@ -1,0 +1,12 @@
+from flask import jsonify
+
+from api import create_app
+
+app = create_app('development')
+
+@app.route('/', methods=['GET'])
+def index():
+    return jsonify(status='API is Up!')
+
+# Import the account related endpoints
+import api.routes.account
