+++ api/database/__init__.py
@@ -3,6 +3,7 @@ from sqlalchemy.inspection import inspect
 
 db = SQLAlchemy()
 
+# From https://stackoverflow.com/a/27951648
 class Serializer(object):
     def serialize(self):
         return {c: getattr(self, c) for c in inspect(self).attrs.keys()}
