+++ api/database/models/__init__.py
@@ -1,0 +1,1 @@
+from api.database.models.account import Account
