--- api/src/database/models/__init__.py
@@ -1,1 +1,0 @@
-from src.database.models.account import Account
