--- api/wsgi.py
@@ -1,5 +1,0 @@
-import os
-
-from src import create_app
-
-app = create_app('development')
